+++ apps/app/playwright/utils/collapse-sidebar.ts
