+++ apps/app/playwright/utils/index.ts
@@ -2,5 +2,5 @@
 // REST API helpers live behind `utils/api` and are imported from there directly;
 // `test-users` is data imported by path. Neither is re-exported here — the barrel
 // exposes only what barrel callers actually use.
-export * from './CollapseSidebar';
-export * from './Login';
+export * from './collapse-sidebar';
+export * from './login';
